Add a global variable for data file
Since the data file is contained in the package contents and the package
scripts are installed and deployed on startup, updating the oui.dat file
requires a redeploy of Zeek or modifying data in the "do-not-touch" folder.
Also, if the package is updated, any local modifications are overwritten.

By using a global variable for the data file path, users can configure
Zeek to use a file outside the package directory, preventing any local
modifications from being overwritten when the package is updated. This
also takes full advantage of the input framework to reread the file on
the fly when changed.